Move `Database` to `database.py`, ensuring plugins.
There's also subexp.py, but don't use it yet, it might be broken for now
